Refactoring moving all module files into one directory
